trw: programs: ignore the `Delete` key presses if there is an active dialog view.

+++ frontend/app/wmes-trw-programs/views/FormView.js
@@ -616,7 +616,7 @@ define([
         this.unselectMsg();
         this.hideEditor();
       }
-      else if (e.key === 'Delete')
+      else if (e.key === 'Delete' && !viewport.currentDialog)
       {
         this.deleteMsg();
       }
